backend/requirements-dev.txt
pytest==8.4.2
python-dotenv==1.0.1
boto3==1.35.0
requests==2.32.0